requirements.txt
torch==2.0.1
torchvision==0.15.2
diffusers==0.25.1
transformers==4.34.0
accelerate==0.23.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.26.0
scikit-image==0.22.0
face-alignment==1.3.5
mediapipe==0.10.8
facexlib==0.3.0
spiga==0.0.6
requests==2.31.0
matplotlib==3.8.0
einops==0.7.0
safetensors==0.3.3
huggingface-hub==0.17.3
git+https://github.com/sajjjadayobi/FaceLib.git@382841efec823f059f49754e5d378abb12cfb551 